refactor: move bridge classes into bridges/base.py; thin __init__ facade
Match the sources/outputs/recipes convention: __init__ is a docstring +
re-export shell, implementation lives in a named module. Import paths
(from myogestic.bridges import Bridge/WebCamBridge/CustomBridge) unchanged.
